[SDXL] Add LoRA to all pipelines
